+++ src/components/stats/DailyApys.tsx
@@ -1,118 +1,165 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-import React, { FC, useMemo } from 'react';
-import { BigNumber, formatUnits } from 'ethers/utils';
-import { VictoryLabel } from 'victory-core';
-import { VictoryLine } from 'victory-line';
-import { VictoryChart } from 'victory-chart';
-import { VictoryAxis } from 'victory-axis';
+import React, { FC } from 'react';
+import {
+  Area,
+  XAxis,
+  YAxis,
+  Tooltip,
+  AreaChart,
+  ResponsiveContainer,
+} from 'recharts';
 import Skeleton from 'react-loading-skeleton';
+import { format } from 'date-fns';
+
 import {
-  endOfDay,
-  fromUnixTime,
-  subDays,
-  startOfDay,
-  getUnixTime,
-  eachDayOfInterval,
-  eachHourOfInterval,
-} from 'date-fns';
-import { useDailApysForGivenTimestamps } from '../../web3/hooks';
+  useDailyApysForBlockTimes,
+  useBlockTimesForDates,
+} from '../../web3/hooks';
 import { Color } from '../../theme';
-import { useVictoryTheme } from '../../context/ThemeProvider';
-import {
-  abbreviateNumber,
-  percentageFormat,
-  useDateFilterTickValues,
-  useDateFilterTickFormat,
-} from './utils';
-import { TimeMetricPeriod } from '../../graphql/legacy';
-import { DateRange } from './Metrics';
+import { percentageFormat, periodFormatMapping } from './utils';
+import { DateRange, Metrics, useDateFilter, useMetrics } from './Metrics';
+import { RechartsContainer } from './RechartsContainer';
 
-const now = new Date();
+enum MetricTypes {
+  DailyApy = 'DailyApy',
+  UtilisationRate = 'UtilisationRate',
+}
 
-const timestamps = eachDayOfInterval({
-  start: subDays(now, 6),
-  end: subDays(now, 1),
-})
-  .map(endOfDay)
-  .concat(now);
+const dailyApyMetrics = [
+  {
+    type: MetricTypes.DailyApy,
+    label: 'Daily APY',
+    color: Color.gold,
+    enabled: true,
+  },
+  {
+    type: MetricTypes.UtilisationRate,
+    label: 'SAVE Utilisation',
+    color: Color.black,
+  },
+];
 
-const dateFilter = {
-  dateRange: DateRange.Week,
-  period: TimeMetricPeriod.Day,
-  label: '7 day',
-  from: timestamps[0],
-  end: timestamps[timestamps.length - 1],
-};
+const formatApy = (percentage: number): string =>
+  percentage > 100
+    ? `${percentageFormat(percentage as number)} 🔥`
+    : percentageFormat(percentage as number);
 
-export const DailyApys: FC<{}> = () => {
-  const dailyApys = useDailApysForGivenTimestamps(timestamps);
-  const tickValues = useDateFilterTickValues(dateFilter);
-  const tickFormat = useDateFilterTickFormat(dateFilter);
-  const victoryTheme = useVictoryTheme();
-  const data = useMemo<{ x: Date; y: number }[]>(
-    () =>
-      dailyApys &&
-      dailyApys
-        .filter(a => a.timestamp && a.apyResults)
-        .map(({ timestamp, apyResults }) => {
-          const percentage = parseFloat(apyResults.dailyAPY);
-          const startTime = fromUnixTime(parseInt(timestamp, 10));
-          return {
-            x: startOfDay(startTime),
-            y: percentage,
-            percentage,
-          };
-        }),
+const DailyApysChart: FC = () => {
+  const dateFilter = useDateFilter();
+  const { DailyApy, UtilisationRate } = useMetrics<MetricTypes>();
 
-    [dailyApys],
-  );
+  const blockTimes = useBlockTimesForDates(dateFilter.dates);
+  const dailyApys = useDailyApysForBlockTimes(blockTimes);
+
   return (
-    <div>
-      {data && data.length ? (
-        <VictoryChart
-          theme={victoryTheme}
-          height={200}
-          domainPadding={25}
-          padding={{ left: 45, top: 10, right: 20, bottom: 40 }}
-        >
-          <VictoryAxis
-            dependentAxis
-            tickFormat={abbreviateNumber}
-            fixLabelOverlap
-            style={{
-              ticks: { stroke: 'none' },
-            }}
-          />
-          <VictoryAxis
-            tickValues={tickValues}
-            tickFormat={tickFormat}
-            style={{
-              grid: { stroke: 'none' },
-            }}
-          />
-          <VictoryLine
-            data={data}
-            labelComponent={<VictoryLabel />}
-            labels={({
-              datum: { percentage },
-            }: {
-              datum: { y: number; percentage: number };
-            }) =>
-              percentage > 100
-                ? `${percentageFormat(percentage)} 🔥`
-                : percentageFormat(percentage)
-            }
-            style={{
-              data: {
-                stroke: Color.gold,
-                strokeWidth: 2,
-              },
-            }}
-          />
-        </VictoryChart>
+    <RechartsContainer>
+      {dailyApys && dailyApys.length ? (
+        <ResponsiveContainer aspect={2}>
+          <AreaChart
+            margin={{ top: 0, right: 16, bottom: 16, left: 16 }}
+            barCategoryGap={1}
+            data={dailyApys}
+          >
+            <defs>
+              <linearGradient id="utilisationRate" x1="0" y1="0" x2="0" y2="1">
+                <stop
+                  offset="5%"
+                  stopColor={Color.blackTransparent}
+                  stopOpacity={0.5}
+                />
+                <stop
+                  offset="95%"
+                  stopColor={Color.blackTransparent}
+                  stopOpacity={0}
+                />
+              </linearGradient>
+              <linearGradient id="dailyAPY" x1="0" y1="0" x2="0" y2="1">
+                <stop offset="5%" stopColor={Color.gold} stopOpacity={0.5} />
+                <stop offset="95%" stopColor={Color.gold} stopOpacity={0} />
+              </linearGradient>
+            </defs>
+            <XAxis
+              dataKey="timestamp"
+              axisLine={false}
+              xAxisId={0}
+              tickSize={12}
+              padding={{ left: 16 }}
+              minTickGap={16}
+              tickLine
+              tickFormatter={timestamp =>
+                timestamp
+                  ? format(
+                      timestamp * 1000,
+                      periodFormatMapping[dateFilter.period],
+                    )
+                  : ''
+              }
+            />
+            <YAxis
+              type="number"
+              orientation="left"
+              tickFormatter={formatApy}
+              axisLine={false}
+              tickLine
+              tickSize={12}
+              padding={{ bottom: 16 }}
+              interval="preserveEnd"
+              minTickGap={8}
+              yAxisId={0}
+            />
+            <Tooltip
+              cursor
+              labelFormatter={timestamp =>
+                format((timestamp as number) * 1000, 'yyyy-MM-dd HH:mm')
+              }
+              formatter={formatApy}
+              separator=""
+              contentStyle={{
+                fontSize: '14px',
+                padding: '8px',
+                background: 'rgba(255, 255, 255, 0.8)',
+                textAlign: 'right',
+                border: 'none',
+                borderRadius: '4px',
+                color: Color.black,
+              }}
+              wrapperStyle={{
+                top: 0,
+                left: 0,
+              }}
+            />
+            <Area
+              hide={!UtilisationRate.enabled}
+              strokeWidth={2}
+              type="monotone"
+              dataKey="utilisationRate"
+              name="SAVE Utilisation "
+              opacity={1}
+              fill="url(#utilisationRate)"
+              yAxisId={0}
+              stroke={Color.blackTransparent}
+            />
+            <Area
+              hide={!DailyApy.enabled}
+              strokeWidth={2}
+              type="monotone"
+              dataKey="dailyAPY"
+              name="Daily APY "
+              opacity={1}
+              yAxisId={0}
+              fill="url(#dailyAPY)"
+              stroke={DailyApy.color}
+            />
+          </AreaChart>
+        </ResponsiveContainer>
       ) : (
         <Skeleton height={270} />
       )}
-    </div>
+    </RechartsContainer>
   );
 };
+
+export const DailyApys: FC = () => (
+  <Metrics defaultDateRange={DateRange.Month} metrics={dailyApyMetrics}>
+    <DailyApysChart />
+  </Metrics>
+);