+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+- Improve analytics page with 30 and 90 day views (including for APY), SAVE utilisation, bAsset vault balances,
+  and fees paid
 - Fetch historic transactions from the subgraph
 - Fetch daily APY from the subgraph